Fixed some more

+++ density/sumxmask.go
@@ -24,7 +24,7 @@ func (sxm *SumXMask) ValueAt(x, y int) (v uint64) {
 		line := sxm.Points[x-sxm.Rect.Min.X]
 		var dv uint64
 		for i := 0; i < len(line) && line[i] < y; i += 2 {
-			dv = line[i+1]
+			dv = uint64(line[i+1])
 		}
 		v = dv
 	}
@@ -39,11 +39,11 @@ func (sxm *SumXMask) ApplyTo(sx *SumX) {
 			var linemass, pv uint64
 			for x := 0; x < sxm.Rect.Dx(); x++ {
 				v := sx.ValueAt(line[x*2], y+sxm.Rect.Min.Y)
-				linemass += (v - pv) * line[x*2+1]
+				linemass += (v - pv) * uint64(line[x*2+1])
 				pv = v
 			}
 			mass += linemass
-			wy += linemass * y
+			wy += linemass * uint64(y)
 		}
 		sxm.Mass = float64(mass) / float64(sxm.Range)
 		sxm.Wy = float64(wy)/float64(mass) + float64(sxm.Rect.Min.Y)
